Agent linux written in bash.
It uses crontab to start/stop nodes and a trivial bash
script to monitor the termination of a node process.

This agent has one limitation comparing to the original implementation:
when computer starts it cannot detect if it's in the middle of period
when nodes must be up. So it will start them on the next cron event.